feat: finish modules/esm
