+++ src/modules/esm.mjs
@@ -1,0 +1,45 @@
+import fs from 'fs';
+import url from 'url';
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+
+const _filename = url.fileURLToPath(import.meta.url);
+const _dirname = path.dirname(_filename);
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = JSON.parse(fs.readFileSync(path.join(_dirname, 'files', 'a.json'), 'utf-8'));
+} else {
+    unknownObject = JSON.parse(fs.readFileSync(path.join(_dirname, 'files', 'b.json'), 'utf-8'));
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${_filename}`);
+console.log(`Path to current directory is ${_dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {
+    unknownObject,
+    myServer,
+};
+
